orca/headers: write wfs.num_sam as float64, not int64
MATLAB conventionally stores numerics as double. With np.int64 the field
loaded as int64 in MATLAB, then data_load_wfs.m line 748 tripped
"Double operands interacting with int64 operands must have integer
values" on
    time = (0:dt:(wfs(wf).Nt_pc-1)*dt).'
because Nt_pc inherited the int64 dtype but dt is double.

Switch to np.float64 so num_sam round-trips as a MATLAB double and the
arithmetic stays in floats. UTIG path (no num_sam) is unaffected.

+++ src/opr_ingest/orca/headers.py
@@ -120,8 +120,12 @@ def get_header_information(rx_samps_path: Union[str, Path]) -> dict:
     # constant 3200 like UTIG MARFA/HICARS. Single-waveform (size-1) struct
     # in MATLAB; `wfs.num_sam == wfs(1).num_sam`, so the OPR indexing pattern
     # `records.settings.wfs(wf).num_sam` works without us having to coax
-    # hdf5storage into emitting a struct array.
-    wfs = {"num_sam": np.array([rx_len_samples], dtype=np.int64)}
+    # hdf5storage into emitting a struct array. dtype=float64 (MATLAB
+    # double) so downstream arithmetic in data_load_wfs.m line 748 -- e.g.
+    # `time = (0:dt:(Nt_pc-1)*dt)` -- doesn't mix int64 with double dt
+    # and trip "Double operands interacting with int64 operands must have
+    # integer values".
+    wfs = {"num_sam": np.array([rx_len_samples], dtype=np.float64)}
 
     return {
         "comp_time": comp_time,
